refactor: funções utilitarias movidas para pasta correta
