+++ src/utils/noteValidation.js
